requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML
anthropic==0.39.0
google-generativeai==0.3.2
openai==1.35.0
numpy==1.24.3
scikit-learn==1.3.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib==1.7.4

# Utils
httpx==0.26.0
aiofiles==23.2.1
python-multipart==0.0.6